프로그래머스 k진수에서 소수 개수 구하기(카카오 블라인드) isPrime 자료형 변환 (long)

+++ src/com/codingtest/exam/programmers/level2/SearchPrimeNumber.java
@@ -38,7 +38,7 @@ public class SearchPrimeNumber {
 
         for (int i = 0; i < list.size(); i++) {
             if(list.get(i).length() > 0) {
-                if(isPrimeNumber(Integer.parseInt(list.get(i)))) {
+                if(isPrimeNumber(Long.parseLong(list.get(i)))) {
                     answer += 1;
                 }
             }
@@ -47,7 +47,7 @@ public class SearchPrimeNumber {
         return answer;
     }
 
-    public boolean isPrimeNumber(int x) {
+    public boolean isPrimeNumber(Long x) {
         if (x == 1) return false;
         int end = (int) Math.sqrt(x);
         for (int i = 2; i <= end; i++) {
